Separate Metrics file; fix under reporting issue
