+++ src/main/java/com/jwplayer/southpaw/metric/StaticGauge.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2018 Longtail Ad Solutions (DBA JW Player)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.jwplayer.southpaw.metric;
+
+import com.codahale.metrics.Gauge;
+
+/**
+ * Simple gauge that we can update manually.
+ * @param <T> - Type stored / reported by the Gauge
+ */
+public class StaticGauge<T> implements Gauge<T> {
+    protected T value;
+
+    public StaticGauge() { }
+
+    @Override
+    public T getValue() {
+        return value;
+    }
+
+    public void update(T value) {
+        this.value = value;
+    }
+}
